Fix segfault when receiving GET request.
The old behaviour is not correct anymore for libpki-0.8.5. In that version
PKI_MEM_url_decode reuses its argument (which is pathmem in this case)
and writes the decoded url at this place. if PKI_MEM_free is called on
pathmem, the decoded url, which is written to b64mem is actually freed
as well. This first leads to use after free and later to a double free
error, which makes the server crash.

Patch was found from http://sourceforge.net/p/openca/mailman/message/32373102/